+++ tests/wps_restapi/test_providers.py
@@ -14,40 +14,20 @@ from tests.utils import (
     setup_mongodb_processstore,
     setup_mongodb_servicestore
 )
+from weaver.config import WEAVER_CONFIGURATION_ADES, WEAVER_CONFIGURATION_HYBRID
+from weaver.datatype import Service
 from weaver.execute import EXECUTE_CONTROL_OPTION_ASYNC, EXECUTE_TRANSMISSION_MODE_REFERENCE
 from weaver.formats import CONTENT_TYPE_APP_JSON, CONTENT_TYPE_APP_NETCDF, CONTENT_TYPE_APP_ZIP, CONTENT_TYPE_TEXT_PLAIN
 from weaver.utils import fully_qualified_name
 
 
 # pylint: disable=C0103,invalid-name
-class WpsRestApiProcessesTest(unittest.TestCase):
-    remote_server = None
-
-    @classmethod
-    def setUpClass(cls):
-        settings = {
-            "weaver.url": "https://localhost",
-            "weaver.wps_path": "/ows/wps",
-        }
-        cls.config = setup_config_with_mongodb(settings=settings)
-        cls.app = get_test_weaver_app(config=cls.config)
-        cls.json_headers = {"Accept": CONTENT_TYPE_APP_JSON, "Content-Type": CONTENT_TYPE_APP_JSON}
-
-    @classmethod
-    def tearDownClass(cls):
-        pyramid.testing.tearDown()
+class WpsProviderBase(unittest.TestCase):
+    remote_provider_name = None
+    settings = {}
 
     def fully_qualified_test_process_name(self):
         return fully_qualified_name(self).replace(".", "-")
-
-    def setUp(self):
-        # rebuild clean db on each test
-        self.service_store = setup_mongodb_servicestore(self.config)
-        self.process_store = setup_mongodb_processstore(self.config)
-        self.job_store = setup_mongodb_jobstore(self.config)
-
-        self.remote_provider_name = "test-remote-provider"
-        self.process_remote_WPS1 = "process_remote_wps1"
 
     def register_provider(self, clear=True, error=False):
         if clear:
@@ -55,8 +35,38 @@ class WpsRestApiProcessesTest(unittest.TestCase):
         path = "/providers"
         data = {"id": self.remote_provider_name, "url": resources.TEST_REMOTE_SERVER_URL}
         resp = self.app.post_json(path, params=data, headers=self.json_headers, expect_errors=error)
-        assert (error and resp.status_code != 201) or (not error and resp.status_code == 201)
+        if error:
+            assert resp.status_code != 201, "Expected provider to fail registration, but erroneously succeeded."
+        else:
+            err = resp.json
+            assert resp.status_code == 201, "Expected failed provider registration to succeed. Error:\n{}".format(err)
         return resp
+
+    @classmethod
+    def setUpClass(cls):
+        cls.config = setup_config_with_mongodb(settings=cls.settings)
+        cls.app = get_test_weaver_app(config=cls.config)
+        cls.json_headers = {"Accept": CONTENT_TYPE_APP_JSON, "Content-Type": CONTENT_TYPE_APP_JSON}
+
+    @classmethod
+    def tearDownClass(cls):
+        pyramid.testing.tearDown()
+
+    def setUp(self):
+        # rebuild clean db on each test
+        self.service_store = setup_mongodb_servicestore(self.config)
+        self.process_store = setup_mongodb_processstore(self.config)
+        self.job_store = setup_mongodb_jobstore(self.config)
+
+
+# pylint: disable=C0103,invalid-name
+class WpsRestApiProvidersTest(WpsProviderBase):
+    remote_provider_name = "test-remote-provider"
+    settings = {
+        "weaver.url": "https://localhost",
+        "weaver.wps_path": "/ows/wps",
+        "weaver.configuration": WEAVER_CONFIGURATION_HYBRID
+    }
 
     @mocked_remote_server_requests_wps1([
         resources.TEST_REMOTE_SERVER_URL,
@@ -438,3 +448,83 @@ class WpsRestApiProcessesTest(unittest.TestCase):
         inputs = resp.json["process"]["inputs"]
         assert "maximumMegabytes" in inputs[11]["formats"][0]
         assert inputs[11]["formats"][0]["maximumMegabytes"] == 200
+
+
+# pylint: disable=C0103,invalid-name
+class WpsProviderLocalOnlyTest(WpsProviderBase):
+    """
+    Validate that operations are preemptively forbidden for a local-only instance.
+    """
+    remote_provider_name = "test-wps-local-only"
+    settings = {
+        "weaver.url": "https://localhost",
+        "weaver.wps_path": "/ows/wps",
+        "weaver.configuration": WEAVER_CONFIGURATION_ADES  # local-only
+    }
+
+    def setUp(self):
+        """
+        Inject service directly in database to ensure errors are not raised by missing entry, but by explicit checks.
+        """
+        super(WpsProviderLocalOnlyTest, self).setUp()
+        self.service_store.clear_services()
+        self.job_store.clear_jobs()
+        self.service_store.save_service(service=Service({
+            "name": self.remote_provider_name,
+            "url": resources.TEST_REMOTE_SERVER_URL,
+            "public": True
+        }), overwrite=True)
+        self.job = self.job_store.save_job("test", "fake-process", self.remote_provider_name)
+
+    @mocked_remote_server_requests_wps1([
+        resources.TEST_REMOTE_SERVER_URL,
+        resources.TEST_REMOTE_PROCESS_GETCAP_WPS1_XML,
+        []
+    ])
+    def test_forbidden_register_provider(self):
+        resp = self.register_provider(error=True)
+        assert resp.status_code == 403, "\n{}".format(resp.json)
+
+    @mocked_remote_server_requests_wps1([
+        resources.TEST_REMOTE_SERVER_URL,
+        resources.TEST_REMOTE_PROCESS_GETCAP_WPS1_XML,
+        [resources.TEST_REMOTE_PROCESS_DESCRIBE_WPS1_XML],
+    ])
+    def test_forbidden_describe_process(self):
+        prov = "/providers/{}".format(self.remote_provider_name)
+        path = "{}/processes/{}/jobs".format(prov, resources.TEST_REMOTE_PROCESS_WPS1_ID)
+        resp = self.app.get(path, headers=self.json_headers, expect_errors=True)
+        assert resp.status_code == 403, "\n{}".format(resp.json)
+
+    @mocked_remote_server_requests_wps1([
+        resources.TEST_REMOTE_SERVER_URL,
+        resources.TEST_REMOTE_PROCESS_GETCAP_WPS1_XML,
+        [resources.TEST_REMOTE_PROCESS_DESCRIBE_WPS1_XML],
+    ])
+    def test_forbidden_execute_process(self):
+        prov = "/providers/{}".format(self.remote_provider_name)
+        path = "{}/processes/{}/jobs".format(prov, resources.TEST_REMOTE_PROCESS_WPS1_ID)
+        resp = self.app.post_json(path, params={}, headers=self.json_headers, expect_errors=True)
+        assert resp.status_code == 403, "\n{}".format(resp.json)
+
+    @mocked_remote_server_requests_wps1([
+        resources.TEST_REMOTE_SERVER_URL,
+        resources.TEST_REMOTE_PROCESS_GETCAP_WPS1_XML,
+        [resources.TEST_REMOTE_PROCESS_DESCRIBE_WPS1_XML],
+    ])
+    def test_forbidden_list_jobs(self):
+        prov = "/providers/{}".format(self.remote_provider_name)
+        path = "{}/processes/{}/jobs".format(prov, resources.TEST_REMOTE_PROCESS_WPS1_ID)
+        resp = self.app.get(path, headers=self.json_headers, expect_errors=True)
+        assert resp.status_code == 403, "\n{}".format(resp.json)
+
+    @mocked_remote_server_requests_wps1([
+        resources.TEST_REMOTE_SERVER_URL,
+        resources.TEST_REMOTE_PROCESS_GETCAP_WPS1_XML,
+        [resources.TEST_REMOTE_PROCESS_DESCRIBE_WPS1_XML],
+    ])
+    def test_forbidden_get_job(self):
+        prov = "/providers/{}".format(self.remote_provider_name)
+        path = "{}/processes/{}/jobs/{}".format(prov, resources.TEST_REMOTE_PROCESS_WPS1_ID, self.job.id)
+        resp = self.app.get(path, headers=self.json_headers, expect_errors=True)
+        assert resp.status_code == 403, "\n{}".format(resp.json)